lessonDOA

+++ module-2/07_Database_Connectivity_DAO/student-exercise/java/src/main/java/com/techelevator/projects/model/Employee.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 public class Employee {
 	// TODO: Place code to describe the columns in the table row here
 	private long employee_id;
-	private int department_id; 
+	private long department_id; 
 	private String first_name;
 	private String last_name;
     private LocalDate birth_date;
@@ -17,10 +17,10 @@ public class Employee {
 	public void setEmployee_id(long employee_id) {
 		this.employee_id = employee_id;
 	}
-	public int getDepartment_id() {
+	public long getDepartment_id() {
 		return department_id;
 	}
-	public void setDepartment_id(int department_id) {
+	public void setDepartment_id(long department_id) {
 		this.department_id = department_id;
 	}
 	public String getFirst_name() {
